requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud APIs
google-generativeai==0.3.2
google-cloud-speech==2.23.0
cloud-sql-python-connector[pg8000]==1.4.3
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0

# Embeddings & Vector DB (FIXED VERSIONS)
sentence-transformers==2.3.1
chromadb==0.4.22
huggingface-hub==0.20.3
numpy<2.0.0
tokenizers>=0.13.3

# Audio Processing
pydub==0.25.1
soundfile==0.12.1

# Telegram Bot
python-telegram-bot==20.7

# Streamlit
streamlit==1.29.0
streamlit-webrtc==0.47.1

# Utils
requests==2.31.0
aiofiles==23.2.1
httpx==0.25.2


plotly==5.18.0